+++ homematicip/class_maps.py
@@ -23,6 +23,7 @@ TYPE_CLASS_MAP = {
     DeviceType.FULL_FLUSH_SHUTTER: FullFlushShutter,
     DeviceType.FULL_FLUSH_SWITCH_MEASURING: FullFlushSwitchMeasuring,
     DeviceType.HEATING_THERMOSTAT: HeatingThermostat,
+    DeviceType.KEY_REMOTE_CONTROL_4: KeyRemoteControl4,
     DeviceType.KEY_REMOTE_CONTROL_ALARM: KeyRemoteControlAlarm,
     DeviceType.MOTION_DETECTOR_INDOOR: MotionDetectorIndoor,
     DeviceType.MOTION_DETECTOR_OUTDOOR: MotionDetectorOutdoor,