+++ tests/test_devices.py
@@ -821,3 +821,8 @@ def test_floor_terminal_block(fake_home: Home):
             "heatingEmergencyValue(0.25) valveProtectionDuration(5) valveProtectionSwitchingInterval(14) pumpFollowUpTime(2) pumpLeadTime(2) "
             "pumpProtectionDuration(1) pumpProtectionSwitchingInterval(14)"
         )
+
+def test_key_remote_control(fake_home:Home):
+    with no_ssl_verification():
+        d = fake_home.search_device_by_id("3014F711ABCDEF0000000014")
+        assert isinstance(d,KeyRemoteControl4)